All tests in a single file avoiding concurrency for jacamo instance
